+++ NEWS.md
@@ -3,6 +3,11 @@
 ---
 
 
+### Bugfixes
+
+- changing error messages
+
+
 ### dependencies
 
 - adding utils